Fix CI failure and repo hygiene
- Skip mnist_smoke in E2E tests (requires data files not available in CI)
- Add .claude/ and CLAUDE.md to .gitignore
- Remove CLAUDE.md from git tracking
- Fix misleading comment in mnist.adam (global max, not per-sample)
- Fix size_t comparison in tensor_load native

+++ vm/src/native.c
@@ -413,7 +413,7 @@ static Value tensor_load_native(VM* vm, int arg_count, Value* args) {
     }
     ObjTensor* tensor = adam_new_tensor(vm, ndim, shape);
     adam_vm_push(vm, OBJ_VAL(tensor)); /* GC protect */
-    if ((int)fread(tensor->data, sizeof(double), count, f) != count) {
+    if (fread(tensor->data, sizeof(double), count, f) != (size_t)count) {
         adam_vm_pop(vm);
         fclose(f);
         return NIL_VAL;
